+++ Console.ShiftsLogger.Ananttuli/ShiftsLoggerApi/Employees/Models/Employee.cs
@@ -7,5 +7,4 @@ public class Employee
     public int EmployeeId { get; set; }
     public required string Name { get; set; }
     public List<Shift> Shifts { get; set; } = [];
-
 }
