+++ Console.ShiftsLogger.Ananttuli/ShiftsLoggerApi/Program.cs
@@ -34,6 +34,5 @@ public class Program
         app.MapControllers();
 
         app.Run();
-
     }
 }
