+++ Console.ShiftsLogger.Ananttuli/ShiftsLoggerApi/Shifts/Models/ShiftCoreDto.cs
@@ -6,15 +6,6 @@ public class ShiftCoreDto
     public DateTime StartTime { get; set; }
     public DateTime EndTime { get; set; }
 
-    public TimeSpan Duration
-    {
-        get
-        {
-            return EndTime.Subtract(StartTime);
-        }
-    }
-
-
     public ShiftCoreDto(
         int shiftId,
         DateTime startTime,
