+++ Console.ShiftsLogger.Ananttuli/ShiftsLoggerUi/Api/ReqUtil.cs
@@ -14,22 +14,6 @@ public class ReqUtil
 
         return item;
     }
-
-    public static string? TryParseJsonStringOrNull(JsonElement element, string propertyName)
-    {
-        bool success = element.TryGetProperty(
-            $"{propertyName}",
-            out JsonElement value
-        );
-
-        if (success)
-        {
-            var cleanedValue = value.ToString().Trim();
-            return cleanedValue.Length > 0 ? cleanedValue : null;
-        }
-
-        return null;
-    }
 }
 
 public class Response<T>(bool success, T data, string? message = null)
