+++ Console.ShiftsLogger.Ananttuli/ShiftsLoggerUi/App/Utils/Text.cs
@@ -6,14 +6,4 @@ public class Text
     {
         return $"[red]{text}[/]";
     }
-
-    public static string Success(string text)
-    {
-        return $"[green]{text}[/]";
-    }
-
-    public static string Markup(string text, string tag)
-    {
-        return $"[{tag}]{text}[/]";
-    }
 }
