+++ Console.ShiftsLogger.Ananttuli/ShiftsLoggerUi/Program.cs
@@ -45,3 +45,19 @@ public class Program
         Console.Clear();
     }
 }
+
+public class MainMenu
+{
+    public const string ManageShifts = "Search & modify all shifts";
+    public const string ManageEmployees = "Manage Employees & Associated Shifts";
+    public const string Exit = "[red]Exit[/]";
+
+    public static string Prompt()
+    {
+        return AnsiConsole.Prompt(
+            new SelectionPrompt<string>()
+                .Title("\nM E N U")
+                .AddChoices([ManageShifts, ManageEmployees, Exit])
+        );
+    }
+}
